RecyclerView with Firestore
Implemented RecyclerView that fetches post details from Firestore (these can easily be adjusted if we need to).

The next thing I will add are OnClick capabilities for each post that brings user to the info page for the NPO, as well as navigation to the main NPO list page (not currently able to access it).